[WIP] create test dir
